+++ vector/src/main/java/im/vector/riotx/features/settings/VectorSettingsAdvancedNotificationPreferenceFragment.kt
@@ -15,125 +15,64 @@
  */
 package im.vector.riotx.features.settings
 
-import android.app.Activity
-import android.content.Intent
-import android.media.RingtoneManager
-import android.net.Uri
-import android.os.Parcelable
-import androidx.core.content.edit
 import androidx.preference.Preference
-import androidx.preference.PreferenceManager
+import im.vector.matrix.android.api.MatrixCallback
+import im.vector.matrix.android.api.pushrules.rest.PushRule
+import im.vector.matrix.android.api.pushrules.rest.PushRuleAndKind
 import im.vector.riotx.R
-import im.vector.riotx.core.preference.BingRule
-import im.vector.riotx.core.preference.BingRulePreference
+import im.vector.riotx.core.preference.PushRulePreference
 import im.vector.riotx.core.preference.VectorPreference
-import im.vector.riotx.features.notifications.NotificationUtils
+import im.vector.riotx.core.utils.toast
 import javax.inject.Inject
 
-class VectorSettingsAdvancedNotificationPreferenceFragment @Inject constructor(
-        private val vectorPreferences: VectorPreferences
-) : VectorSettingsBaseFragment() {
-
-    // events listener
-    /* TODO
-    private val mEventsListener = object : MXEventListener() {
-        override fun onBingRulesUpdate() {
-            refreshPreferences()
-            refreshDisplay()
-        }
-    }    */
+class VectorSettingsAdvancedNotificationPreferenceFragment @Inject constructor()
+    : VectorSettingsBaseFragment() {
 
     override var titleRes: Int = R.string.settings_notification_advanced
 
     override val preferenceXmlRes = R.xml.vector_settings_notification_advanced_preferences
 
     override fun bindPref() {
-        val callNotificationsSystemOptions = findPreference<VectorPreference>(VectorPreferences.SETTINGS_SYSTEM_CALL_NOTIFICATION_PREFERENCE_KEY)!!
-        if (NotificationUtils.supportNotificationChannels()) {
-            callNotificationsSystemOptions.onPreferenceClickListener = Preference.OnPreferenceClickListener {
-                NotificationUtils.openSystemSettingsForCallCategory(this)
-                false
-            }
-        } else {
-            callNotificationsSystemOptions.isVisible = false
-        }
-
-        val noisyNotificationsSystemOptions = findPreference<VectorPreference>(VectorPreferences.SETTINGS_SYSTEM_NOISY_NOTIFICATION_PREFERENCE_KEY)!!
-        if (NotificationUtils.supportNotificationChannels()) {
-            noisyNotificationsSystemOptions.onPreferenceClickListener = Preference.OnPreferenceClickListener {
-                NotificationUtils.openSystemSettingsForNoisyCategory(this)
-                false
-            }
-        } else {
-            noisyNotificationsSystemOptions.isVisible = false
-        }
-
-        val silentNotificationsSystemOptions = findPreference<VectorPreference>(VectorPreferences.SETTINGS_SYSTEM_SILENT_NOTIFICATION_PREFERENCE_KEY)!!
-        if (NotificationUtils.supportNotificationChannels()) {
-            silentNotificationsSystemOptions.onPreferenceClickListener = Preference.OnPreferenceClickListener {
-                NotificationUtils.openSystemSettingsForSilentCategory(this)
-                false
-            }
-        } else {
-            silentNotificationsSystemOptions.isVisible = false
-        }
-
-        // Ringtone
-        val ringtonePreference = findPreference<VectorPreference>(VectorPreferences.SETTINGS_NOTIFICATION_RINGTONE_SELECTION_PREFERENCE_KEY)!!
-
-        if (NotificationUtils.supportNotificationChannels()) {
-            ringtonePreference.isVisible = false
-        } else {
-            ringtonePreference.summary = vectorPreferences.getNotificationRingToneName()
-            ringtonePreference.onPreferenceClickListener = Preference.OnPreferenceClickListener {
-                val intent = Intent(RingtoneManager.ACTION_RINGTONE_PICKER)
-                intent.putExtra(RingtoneManager.EXTRA_RINGTONE_TYPE, RingtoneManager.TYPE_NOTIFICATION)
-
-                if (null != vectorPreferences.getNotificationRingTone()) {
-                    intent.putExtra(RingtoneManager.EXTRA_RINGTONE_EXISTING_URI, vectorPreferences.getNotificationRingTone())
-                }
-
-                startActivityForResult(intent, REQUEST_NOTIFICATION_RINGTONE)
-                false
-            }
-        }
-
-        for (preferenceKey in mPrefKeyToBingRuleId.keys) {
+        for (preferenceKey in prefKeyToPushRuleId.keys) {
             val preference = findPreference<VectorPreference>(preferenceKey)
-            if (preference is BingRulePreference) {
+            if (preference is PushRulePreference) {
                 // preference.isEnabled = null != rules && isConnected && pushManager.areDeviceNotificationsAllowed()
-                val rule: BingRule? = null // TODO session.dataHandler.pushRules()?.findDefaultRule(mPrefKeyToBingRuleId[preferenceKey])
+                val ruleAndKind: PushRuleAndKind? = session.getPushRules().findDefaultRule(prefKeyToPushRuleId[preferenceKey])
 
-                if (rule == null) {
+                if (ruleAndKind == null) {
                     // The rule is not defined, hide the preference
                     preference.isVisible = false
                 } else {
                     preference.isVisible = true
-                    preference.setBingRule(rule)
+                    preference.setPushRule(ruleAndKind)
                     preference.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _, newValue ->
-                        val rule2 = preference.createRule(newValue as Int)
-                        if (null != rule2) {
-                            /*
-                              TODO
-                           displayLoadingView()
-                            session.dataHandler.bingRulesManager.updateRule(preference.rule,
-                                    rule,
-                                    object : BingRulesManager.onBingRuleUpdateListener {
-                                        private fun onDone() {
-                                            refreshDisplay()
+                        val newRule = preference.createNewRule(newValue as Int)
+                        if (newRule != null) {
+                            displayLoadingView()
+
+                            session.updatePushRuleActions(
+                                    ruleAndKind.kind,
+                                    preference.ruleAndKind?.pushRule ?: ruleAndKind.pushRule,
+                                    newRule,
+                                    object : MatrixCallback<Unit> {
+                                        override fun onSuccess(data: Unit) {
+                                            if (!isAdded) {
+                                                return
+                                            }
+                                            preference.setPushRule(ruleAndKind.copy(pushRule = newRule))
                                             hideLoadingView()
                                         }
 
-                                        override fun onBingRuleUpdateSuccess() {
-                                            onDone()
-                                        }
-
-                                        override fun onBingRuleUpdateFailure(errorMessage: String) {
-                                            activity?.toast(errorMessage)
-                                            onDone()
+                                        override fun onFailure(failure: Throwable) {
+                                            if (!isAdded) {
+                                                return
+                                            }
+                                            hideLoadingView()
+                                            // Restore the previous value
+                                            refreshDisplay()
+                                            activity?.toast(errorFormatter.toHumanReadable(failure))
                                         }
                                     })
-                                    */
                         }
                         false
                     }
@@ -146,83 +85,24 @@ class VectorSettingsAdvancedNotificationPreferenceFragment @Inject constructor(
         listView?.adapter?.notifyDataSetChanged()
     }
 
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (resultCode == Activity.RESULT_OK) {
-            when (requestCode) {
-                REQUEST_NOTIFICATION_RINGTONE -> {
-                    vectorPreferences.setNotificationRingTone(data?.getParcelableExtra<Parcelable>(RingtoneManager.EXTRA_RINGTONE_PICKED_URI) as Uri?)
-
-                    // test if the selected ring tone can be played
-                    val notificationRingToneName = vectorPreferences.getNotificationRingToneName()
-                    if (null != notificationRingToneName) {
-                        vectorPreferences.setNotificationRingTone(vectorPreferences.getNotificationRingTone())
-                        findPreference<VectorPreference>(VectorPreferences.SETTINGS_NOTIFICATION_RINGTONE_SELECTION_PREFERENCE_KEY)!!
-                                .summary = notificationRingToneName
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Refresh the known information about the account
-     */
-    private fun refreshPreferences() {
-        PreferenceManager.getDefaultSharedPreferences(activity).edit {
-            /* TODO
-            session.dataHandler.pushRules()?.let {
-                for (prefKey in mPrefKeyToBingRuleId.keys) {
-                    val preference = findPreference(prefKey)
-
-                    if (null != preference && preference is SwitchPreference) {
-                        val ruleId = mPrefKeyToBingRuleId[prefKey]
-
-                        val rule = it.findDefaultRule(ruleId)
-                        var isEnabled = null != rule && rule.isEnabled
-
-                        if (TextUtils.equals(ruleId, BingRule.RULE_ID_DISABLE_ALL) || TextUtils.equals(ruleId, BingRule.RULE_ID_SUPPRESS_BOTS_NOTIFICATIONS)) {
-                            isEnabled = !isEnabled
-                        } else if (isEnabled) {
-                            val domainActions = rule!!.domainActions
-
-                            // no action -> noting will be done
-                            if (null == domainActions || domainActions.isEmpty()) {
-                                isEnabled = false
-                            } else if (1 == domainActions.size) {
-                                try {
-                                    isEnabled = !TextUtils.equals(domainActions[0] as String, BingRule.ACTION_DONT_NOTIFY)
-                                } catch (e: Exception) {
-                                    Timber.e(e, "## refreshPreferences failed")
-                                }
-
-                            }
-                        }// check if the rule is only defined by don't notify
-
-                        putBoolean(prefKey, isEnabled)
-                    }
-                }
-            }
-            */
-        }
-    }
-
     /* ==========================================================================================
      * Companion
      * ========================================================================================== */
 
     companion object {
-        private const val REQUEST_NOTIFICATION_RINGTONE = 888
-
         //  preference name <-> rule Id
-        private var mPrefKeyToBingRuleId = mapOf(
-                VectorPreferences.SETTINGS_CONTAINING_MY_DISPLAY_NAME_PREFERENCE_KEY to BingRule.RULE_ID_CONTAIN_DISPLAY_NAME,
-                VectorPreferences.SETTINGS_CONTAINING_MY_USER_NAME_PREFERENCE_KEY to BingRule.RULE_ID_CONTAIN_USER_NAME,
-                VectorPreferences.SETTINGS_MESSAGES_IN_ONE_TO_ONE_PREFERENCE_KEY to BingRule.RULE_ID_ONE_TO_ONE_ROOM,
-                VectorPreferences.SETTINGS_MESSAGES_IN_GROUP_CHAT_PREFERENCE_KEY to BingRule.RULE_ID_ALL_OTHER_MESSAGES_ROOMS,
-                VectorPreferences.SETTINGS_INVITED_TO_ROOM_PREFERENCE_KEY to BingRule.RULE_ID_INVITE_ME,
-                VectorPreferences.SETTINGS_CALL_INVITATIONS_PREFERENCE_KEY to BingRule.RULE_ID_CALL,
-                VectorPreferences.SETTINGS_MESSAGES_SENT_BY_BOT_PREFERENCE_KEY to BingRule.RULE_ID_SUPPRESS_BOTS_NOTIFICATIONS
+        private val prefKeyToPushRuleId = mapOf(
+                "SETTINGS_PUSH_RULE_CONTAINING_MY_DISPLAY_NAME_PREFERENCE_KEY" to PushRule.RULE_ID_CONTAIN_DISPLAY_NAME,
+                "SETTINGS_PUSH_RULE_CONTAINING_MY_USER_NAME_PREFERENCE_KEY" to PushRule.RULE_ID_CONTAIN_USER_NAME,
+                "SETTINGS_PUSH_RULE_MESSAGES_IN_ONE_TO_ONE_PREFERENCE_KEY" to PushRule.RULE_ID_ONE_TO_ONE_ROOM,
+                "SETTINGS_PUSH_RULE_MESSAGES_IN_GROUP_CHAT_PREFERENCE_KEY" to PushRule.RULE_ID_ALL_OTHER_MESSAGES_ROOMS,
+                "SETTINGS_PUSH_RULE_INVITED_TO_ROOM_PREFERENCE_KEY" to PushRule.RULE_ID_INVITE_ME,
+                "SETTINGS_PUSH_RULE_CALL_INVITATIONS_PREFERENCE_KEY" to PushRule.RULE_ID_CALL,
+                "SETTINGS_PUSH_RULE_MESSAGES_SENT_BY_BOT_PREFERENCE_KEY" to PushRule.RULE_ID_SUPPRESS_BOTS_NOTIFICATIONS,
+                "SETTINGS_PUSH_RULE_MESSAGES_CONTAINING_AT_ROOM_PREFERENCE_KEY" to PushRule.RULE_ID_AT_ROOMS,
+                "SETTINGS_PUSH_RULE_MESSAGES_IN_E2E_ONE_ONE_CHAT_PREFERENCE_KEY" to PushRule.RULE_ID_E2E_ONE_TO_ONE_ROOM,
+                "SETTINGS_PUSH_RULE_MESSAGES_IN_E2E_GROUP_CHAT_PREFERENCE_KEY" to PushRule.RULE_ID_E2E_GROUP,
+                "SETTINGS_PUSH_RULE_ROOMS_UPGRADED_KEY" to PushRule.RULE_ID_TOMBSTONE
         )
     }
 }